vmaas_sync configure method moved to main component source file
